Fully cleaned gruntfile.js
